+++ src/modules/rolle/domain/apply-rollenerweiterung-role-angebote-service.spec.ts
@@ -3,15 +3,12 @@ import { OrganisationRepository } from '../../organisation/persistence/organisat
 import { RolleRepo } from '../repo/rolle.repo.js';
 import { RollenerweiterungRepo } from '../repo/rollenerweiterung.repo.js';
 import { PersonPermissions } from '../../authentication/domain/person-permissions.js';
-import { ApplyRollenerweiterungBodyParams } from '../api/apply-rollenerweiterung.body.params.js';
 import { Rollenerweiterung } from './rollenerweiterung.js';
 import { Rolle } from './rolle.js';
 import { Ok } from '../../../shared/util/result.js';
 import { EntityNotFoundError } from '../../../shared/error/entity-not-found.error.js';
-import { ApplyRollenerweiterungRolesError } from '../api/apply-rollenerweiterung-roles.error.js';
 import { faker } from '@faker-js/faker';
 import { createMock, DeepMocked } from '../../../../test/utils/createMock.js';
-import { ApplyRollenerweiterungService } from './apply-rollenerweiterungen-service.js';
 import { MissingPermissionsError } from '../../../shared/error/index.js';
 import { MissingMerkmalVerfuegbarFuerRollenerweiterungError } from './missing-merkmal-verfuegbar-fuer-rollenerweiterung.error.js';
 import { DoFactory } from '../../../../test/utils/do-factory.js';
@@ -25,21 +22,14 @@ import { ServiceProviderMerkmal } from '../../service-provider/domain/service-pr
 import { ApplyRollenerweiterungServiceProvidersError } from '../api/apply-rollenerweiterung-service-providers.error.js';
 import { ServiceProvider } from '../../service-provider/domain/service-provider.js';
 import { ApplyRollenerweiterungChangesBodyParams } from '../api/apply-rollenerweiterung-changes.body.params.js';
+import { ApplyRollenerweiterungWithRoleForAngeboteService } from './apply-rollenerweiterungen-role-angebote-service.js';
 
-type TresultType = Result<
-    null,
-    | ApplyRollenerweiterungRolesError
-    | EntityNotFoundError
-    | MissingPermissionsError
-    | MissingMerkmalVerfuegbarFuerRollenerweiterungError
->;
-
-describe('ApplyRollenerweiterungService', () => {
+describe('ApplyRollenerweiterungWithRoleForAngeboteService', () => {
     let serviceProviderRepo: DeepMocked<ServiceProviderRepo>;
     let organisationRepo: DeepMocked<OrganisationRepository>;
     let rolleRepo: DeepMocked<RolleRepo>;
     let rollenerweiterungRepo: DeepMocked<RollenerweiterungRepo>;
-    let service: ApplyRollenerweiterungService;
+    let service: ApplyRollenerweiterungWithRoleForAngeboteService;
 
     beforeAll(async () => {
         const module: TestingModule = await Test.createTestingModule({
@@ -61,7 +51,7 @@ describe('ApplyRollenerweiterungService', () => {
                     provide: RolleRepo,
                     useValue: createMock(RolleRepo),
                 },
-                ApplyRollenerweiterungService,
+                ApplyRollenerweiterungWithRoleForAngeboteService,
             ],
         }).compile();
 
@@ -69,255 +59,11 @@ describe('ApplyRollenerweiterungService', () => {
         organisationRepo = module.get(OrganisationRepository);
         rolleRepo = module.get(RolleRepo);
         rollenerweiterungRepo = module.get(RollenerweiterungRepo);
-        service = module.get(ApplyRollenerweiterungService);
+        service = module.get(ApplyRollenerweiterungWithRoleForAngeboteService);
     }, DEFAULT_TIMEOUT_FOR_TESTCONTAINERS);
 
     beforeEach(() => {
         vi.resetAllMocks();
-    });
-
-    it('should add and remove Erweiterungen successfully', async () => {
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-        const rolleIdAdd: string = faker.string.uuid();
-        const rolleIdRemove: string = faker.string.uuid();
-
-        organisationRepo.findById.mockResolvedValue(DoFactory.createOrganisation(true, { id: orgaId }));
-        serviceProviderRepo.findById.mockResolvedValue(
-            DoFactory.createServiceProvider(true, {
-                id: angebotId,
-                merkmale: [ServiceProviderMerkmal.VERFUEGBAR_FUER_ROLLENERWEITERUNG],
-            }),
-        );
-
-        const existingErw: Rollenerweiterung<true> = createMock<Rollenerweiterung<true>>(Rollenerweiterung, {
-            rolleId: rolleIdRemove,
-        });
-        rollenerweiterungRepo.findManyByOrganisationIdAndServiceProviderId.mockResolvedValue([existingErw]);
-
-        const rolleAdd: Rolle<true> = createMock<Rolle<true>>(Rolle);
-        const rolleRemove: Rolle<true> = createMock<Rolle<true>>(Rolle);
-        rolleRepo.findByIds.mockResolvedValue(
-            new Map([
-                [rolleIdAdd, rolleAdd],
-                [rolleIdRemove, rolleRemove],
-            ]),
-        );
-
-        rollenerweiterungRepo.createAuthorized.mockResolvedValue(
-            Ok(createMock<Rollenerweiterung<true>>(Rollenerweiterung)),
-        );
-        rollenerweiterungRepo.deleteByComposedId.mockResolvedValue(Ok(null));
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [rolleIdAdd],
-            removeErweiterungenForRolleIds: [rolleIdRemove],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(true);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(true);
-        expect(rollenerweiterungRepo.createAuthorized).toHaveBeenCalled();
-        expect(rollenerweiterungRepo.deleteByComposedId).toHaveBeenCalled();
-    });
-
-    it('should return error if Permissions are missing', async () => {
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-
-        organisationRepo.findById.mockResolvedValue(DoFactory.createOrganisation(true, { id: orgaId }));
-        serviceProviderRepo.findById.mockResolvedValue(
-            DoFactory.createServiceProvider(true, {
-                id: angebotId,
-                merkmale: [ServiceProviderMerkmal.VERFUEGBAR_FUER_ROLLENERWEITERUNG],
-            }),
-        );
-        rolleRepo.findByIds.mockResolvedValue(new Map());
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [],
-            removeErweiterungenForRolleIds: [],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(false);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(false);
-        if (result.ok) {
-            return;
-        }
-        expect(result.error).toBeInstanceOf(MissingPermissionsError);
-        const err: unknown = result.error;
-        if (!(err instanceof MissingPermissionsError)) {
-            return;
-        }
-    });
-
-    it('should return error if Organisation is missing', async () => {
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-
-        organisationRepo.findById.mockResolvedValue(undefined);
-        serviceProviderRepo.findById.mockResolvedValue(
-            DoFactory.createServiceProvider(true, {
-                id: angebotId,
-                merkmale: [ServiceProviderMerkmal.VERFUEGBAR_FUER_ROLLENERWEITERUNG],
-            }),
-        );
-        rolleRepo.findByIds.mockResolvedValue(new Map());
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [],
-            removeErweiterungenForRolleIds: [],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(true);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(false);
-        if (result.ok) {
-            return;
-        }
-        expect(result.error).toBeInstanceOf(EntityNotFoundError);
-        const err: unknown = result.error;
-        if (!(err instanceof EntityNotFoundError)) {
-            return;
-        }
-    });
-
-    it('should return error if Angebot is missing', async () => {
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-
-        organisationRepo.findById.mockResolvedValue(DoFactory.createOrganisation(true, { id: orgaId }));
-        serviceProviderRepo.findById.mockResolvedValue(undefined);
-        rolleRepo.findByIds.mockResolvedValue(new Map());
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [],
-            removeErweiterungenForRolleIds: [],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(true);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(false);
-        if (result.ok) {
-            return;
-        }
-        expect(result.error).toBeInstanceOf(EntityNotFoundError);
-        const err: unknown = result.error;
-        if (!(err instanceof EntityNotFoundError)) {
-            return;
-        }
-    });
-
-    it('should return error if Angebot is not erweiterbar', async () => {
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-
-        organisationRepo.findById.mockResolvedValue(DoFactory.createOrganisation(true, { id: orgaId }));
-        serviceProviderRepo.findById.mockResolvedValue(
-            DoFactory.createServiceProvider(true, { id: angebotId, merkmale: [] }),
-        );
-        rolleRepo.findByIds.mockResolvedValue(new Map());
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [],
-            removeErweiterungenForRolleIds: [],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(true);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(false);
-        if (result.ok) {
-            return;
-        }
-        expect(result.error).toBeInstanceOf(MissingMerkmalVerfuegbarFuerRollenerweiterungError);
-        const err: unknown = result.error;
-        if (!(err instanceof MissingMerkmalVerfuegbarFuerRollenerweiterungError)) {
-            return;
-        }
-    });
-
-    it('should return error if Rolle not found for add', async () => {
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-        const rolleId: string = faker.string.uuid();
-
-        rollenerweiterungRepo.findManyByOrganisationIdAndServiceProviderId.mockResolvedValue([]);
-        rolleRepo.findByIds.mockResolvedValue(new Map());
-
-        organisationRepo.findById.mockResolvedValue(DoFactory.createOrganisation(true, { id: orgaId }));
-        serviceProviderRepo.findById.mockResolvedValue(
-            DoFactory.createServiceProvider(true, {
-                id: angebotId,
-                merkmale: [ServiceProviderMerkmal.VERFUEGBAR_FUER_ROLLENERWEITERUNG],
-            }),
-        );
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [rolleId],
-            removeErweiterungenForRolleIds: [],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(true);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(false);
-        if (result.ok) {
-            return;
-        }
-        expect(result.error).toBeInstanceOf(ApplyRollenerweiterungRolesError);
-        const err: unknown = result.error;
-        if (!(err instanceof ApplyRollenerweiterungRolesError)) {
-            return;
-        }
-        expect(err.errors[0]?.id).toBe(rolleId);
-        expect(err.errors[0]?.error).toBeInstanceOf(EntityNotFoundError);
-    });
-
-    it('should return error if Rolle not found for remove', async () => {
-        const rolleIdRemove: string = faker.string.uuid();
-
-        const orgaId: string = faker.string.uuid();
-        const angebotId: string = faker.string.uuid();
-
-        organisationRepo.findById.mockResolvedValue(DoFactory.createOrganisation(true, { id: orgaId }));
-        serviceProviderRepo.findById.mockResolvedValue(
-            DoFactory.createServiceProvider(true, {
-                id: angebotId,
-                merkmale: [ServiceProviderMerkmal.VERFUEGBAR_FUER_ROLLENERWEITERUNG],
-            }),
-        );
-
-        const existingErw: Rollenerweiterung<true> = createMock<Rollenerweiterung<true>>(Rollenerweiterung, {
-            rolleId: rolleIdRemove,
-        });
-        rollenerweiterungRepo.findManyByOrganisationIdAndServiceProviderId.mockResolvedValue([existingErw]);
-        rolleRepo.findByIds.mockResolvedValue(new Map());
-
-        const body: ApplyRollenerweiterungBodyParams = {
-            addErweiterungenForRolleIds: [],
-            removeErweiterungenForRolleIds: [rolleIdRemove],
-        };
-        const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
-        permissions.hasSystemrechtAtOrganisation.mockResolvedValueOnce(true);
-
-        const result: TresultType = await service.applyRollenerweiterungChanges(orgaId, angebotId, body, permissions);
-        expect(result.ok).toBe(false);
-        if (result.ok) {
-            return;
-        }
-        expect(result.error).toBeInstanceOf(ApplyRollenerweiterungRolesError);
-        const err: unknown = result.error;
-        if (!(err instanceof ApplyRollenerweiterungRolesError)) {
-            return;
-        }
-        expect(err.errors[0]?.id).toBe(rolleIdRemove);
-        expect(err.errors[0]?.error).toBeInstanceOf(EntityNotFoundError);
     });
 
     describe('applyRollenerweiterungChangesForRolle', () => {
@@ -373,7 +119,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(true);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
@@ -396,7 +142,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(false);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
@@ -423,7 +169,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(true);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
@@ -451,7 +197,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(true);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
@@ -485,7 +231,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(true);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
@@ -532,7 +278,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(true);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
@@ -581,7 +327,7 @@ describe('ApplyRollenerweiterungService', () => {
             const permissions: DeepMocked<PersonPermissions> = createPersonPermissionsMock();
             permissions.hasSystemrechtAtOrganisation.mockResolvedValue(true);
 
-            const result: TresultType = await service.applyRollenerweiterungChangesForRolle(
+            const result: TresultType = await service.applyRollenerweiterungChangesWithRoleForAngebote(
                 orgaId,
                 rolleId,
                 body,
